Rename Serializer class to BaseSerializer
