+++ src/lib/bson/json.ts
@@ -1,6 +1,5 @@
 import { Incident } from "incident";
-import { Serializer } from "../serializer";
-import { TypeSerializer } from "../types";
+import { Serializer, TypeSerializer } from "../types";
 import { JsonType, name as typeName } from "../types/json";
 
 function write(type: JsonType, value: any): any {
