+++ src/lib/qs/codepoint-string.ts
@@ -1,5 +1,4 @@
-import { Serializer } from "../serializer";
-import { TypeSerializer } from "../types";
+import { Serializer, TypeSerializer } from "../types";
 import { CodepointStringType, name as typeName } from "../types/codepoint-string";
 
 function write(type: CodepointStringType, value: string): string {
