+++ src/lib/qs/float64.ts
@@ -1,7 +1,6 @@
 import { Incident } from "incident";
 import { createInvalidTypeError } from "../errors/invalid-type";
-import { Serializer } from "../serializer";
-import { TypeSerializer } from "../types";
+import { Serializer, TypeSerializer } from "../types";
 import { Float64Type, name as typeName } from "../types/float64";
 
 function write(type: Float64Type, value: number): string {
